test: extract MockCommandRepository
