+++ internal/command/application/handlers/cleancommandsonprojectdeleted_test.go
@@ -5,49 +5,11 @@ import (
 	"testing"
 
 	"github.com/stretchr/testify/assert"
-	"github.com/stretchr/testify/mock"
 
 	"gomander/internal/command/application/handlers"
-	commanddomain "gomander/internal/command/domain"
+	"gomander/internal/command/domain/test"
 	projectdomainevent "gomander/internal/project/domain/event"
 )
-
-type MockCommandRepository struct {
-	mock.Mock
-}
-
-func (m *MockCommandRepository) Get(commandId string) (*commanddomain.Command, error) {
-	args := m.Called(commandId)
-	if args.Get(0) == nil {
-		return nil, args.Error(1)
-	}
-	return args.Get(0).(*commanddomain.Command), args.Error(1)
-}
-
-func (m *MockCommandRepository) GetAll(projectId string) ([]commanddomain.Command, error) {
-	args := m.Called(projectId)
-	return args.Get(0).([]commanddomain.Command), args.Error(1)
-}
-
-func (m *MockCommandRepository) Create(command *commanddomain.Command) error {
-	args := m.Called(command)
-	return args.Error(0)
-}
-
-func (m *MockCommandRepository) Update(command *commanddomain.Command) error {
-	args := m.Called(command)
-	return args.Error(0)
-}
-
-func (m *MockCommandRepository) Delete(commandId string) error {
-	args := m.Called(commandId)
-	return args.Error(0)
-}
-
-func (m *MockCommandRepository) DeleteAll(projectId string) error {
-	args := m.Called(projectId)
-	return args.Error(0)
-}
 
 type FakeEvent struct{}
 
@@ -58,7 +20,7 @@ var pjId = "pj-123"
 func TestDefaultCleanCommandsOnProjectDeleted(t *testing.T) {
 	t.Run("Should remove command from command groups and delete empty groups", func(t *testing.T) {
 		// Arrange
-		mockRepo := new(MockCommandRepository)
+		mockRepo := new(test.MockCommandRepository)
 		handler := handlers.NewCleanCommandOnProjectDeleted(mockRepo)
 		event := projectdomainevent.ProjectDeletedEvent{ProjectId: pjId}
 
@@ -74,7 +36,7 @@ func TestDefaultCleanCommandsOnProjectDeleted(t *testing.T) {
 
 	t.Run("Should return error if failing to remove command from command groups", func(t *testing.T) {
 		// Arrange
-		mockRepo := new(MockCommandRepository)
+		mockRepo := new(test.MockCommandRepository)
 		handler := handlers.NewCleanCommandOnProjectDeleted(mockRepo)
 		event := projectdomainevent.ProjectDeletedEvent{ProjectId: pjId}
 
@@ -91,7 +53,7 @@ func TestDefaultCleanCommandsOnProjectDeleted(t *testing.T) {
 
 	t.Run("Should do nothing if command is the wrong type", func(t *testing.T) {
 		// Arrange
-		mockRepo := new(MockCommandRepository)
+		mockRepo := new(test.MockCommandRepository)
 		handler := handlers.NewCleanCommandOnProjectDeleted(mockRepo)
 
 		// Act
