+++ internal/command/application/usecases/removecommand_test.go
@@ -9,12 +9,13 @@ import (
 
 	"gomander/internal/command/application/usecases"
 	commanddomainevent "gomander/internal/command/domain/event"
+	"gomander/internal/command/domain/test"
 )
 
 func TestDefaultRemoveCommand_Execute(t *testing.T) {
 	t.Run("Should remove the command", func(t *testing.T) {
 		// Arrange
-		mockCommandRepository := new(MockCommandRepository)
+		mockCommandRepository := new(test.MockCommandRepository)
 		mockEventBus := new(MockEventBus)
 
 		sut := usecases.NewRemoveCommand(mockCommandRepository, mockEventBus)
@@ -38,7 +39,7 @@ func TestDefaultRemoveCommand_Execute(t *testing.T) {
 
 	t.Run("Should return an error if fails to remove the command", func(t *testing.T) {
 		// Arrange
-		mockCommandRepository := new(MockCommandRepository)
+		mockCommandRepository := new(test.MockCommandRepository)
 		mockEventBus := new(MockEventBus)
 
 		sut := usecases.NewRemoveCommand(mockCommandRepository, mockEventBus)
@@ -60,7 +61,7 @@ func TestDefaultRemoveCommand_Execute(t *testing.T) {
 
 	t.Run("Should return an error if side effect fail", func(t *testing.T) {
 		// Arrange
-		mockCommandRepository := new(MockCommandRepository)
+		mockCommandRepository := new(test.MockCommandRepository)
 		mockEventBus := new(MockEventBus)
 
 		sut := usecases.NewRemoveCommand(mockCommandRepository, mockEventBus)
