+++ internal/commandgroup/application/handlers/addcommandtogrouponcommandduplicated_test.go
@@ -28,7 +28,7 @@ func TestDefaultAddCommandToGroupOnCommandDuplicated_GetEvent(t *testing.T) {
 func TestDefaultAddCommandToGroupOnCommandDuplicated(t *testing.T) {
 	t.Run("Should do nothing if command was not duplicated inside a group", func(t *testing.T) {
 		// Arrange
-		mockCommandRepo := new(MockCommandRepository)
+		mockCommandRepo := new(test.MockCommandRepository)
 		mockCommandGroupRepo := new(MockCommandGroupRepository)
 		handler := handlers.NewAddCommandToGroupOnCommandDuplicated(mockCommandRepo, mockCommandGroupRepo)
 		event := commanddomainevent.CommandDuplicatedEvent{
@@ -47,7 +47,7 @@ func TestDefaultAddCommandToGroupOnCommandDuplicated(t *testing.T) {
 
 	t.Run("Should add duplicated command to the group", func(t *testing.T) {
 		// Arrange
-		mockCommandRepo := new(MockCommandRepository)
+		mockCommandRepo := new(test.MockCommandRepository)
 		mockCommandGroupRepo := new(MockCommandGroupRepository)
 		handler := handlers.NewAddCommandToGroupOnCommandDuplicated(mockCommandRepo, mockCommandGroupRepo)
 		event := commanddomainevent.CommandDuplicatedEvent{
@@ -76,7 +76,7 @@ func TestDefaultAddCommandToGroupOnCommandDuplicated(t *testing.T) {
 
 	t.Run("Should do nothing if command is already in the group", func(t *testing.T) {
 		// Arrange
-		mockCommandRepo := new(MockCommandRepository)
+		mockCommandRepo := new(test.MockCommandRepository)
 		mockCommandGroupRepo := new(MockCommandGroupRepository)
 		handler := handlers.NewAddCommandToGroupOnCommandDuplicated(mockCommandRepo, mockCommandGroupRepo)
 		event := commanddomainevent.CommandDuplicatedEvent{
@@ -101,7 +101,7 @@ func TestDefaultAddCommandToGroupOnCommandDuplicated(t *testing.T) {
 
 	t.Run("Should do nothing if command is the wrong type", func(t *testing.T) {
 		// Arrange
-		mockCommandRepo := new(MockCommandRepository)
+		mockCommandRepo := new(test.MockCommandRepository)
 		mockCommandGroupRepo := new(MockCommandGroupRepository)
 		handler := handlers.NewAddCommandToGroupOnCommandDuplicated(mockCommandRepo, mockCommandGroupRepo)
 
@@ -115,7 +115,7 @@ func TestDefaultAddCommandToGroupOnCommandDuplicated(t *testing.T) {
 
 	t.Run("Should return error if failing to get command group", func(t *testing.T) {
 		// Arrange
-		mockCommandRepo := new(MockCommandRepository)
+		mockCommandRepo := new(test.MockCommandRepository)
 		mockCommandGroupRepo := new(MockCommandGroupRepository)
 		handler := handlers.NewAddCommandToGroupOnCommandDuplicated(mockCommandRepo, mockCommandGroupRepo)
 		event := commanddomainevent.CommandDuplicatedEvent{
@@ -136,7 +136,7 @@ func TestDefaultAddCommandToGroupOnCommandDuplicated(t *testing.T) {
 
 	t.Run("Should return error if failing to get duplicated command", func(t *testing.T) {
 		// Arrange
-		mockCommandRepo := new(MockCommandRepository)
+		mockCommandRepo := new(test.MockCommandRepository)
 		mockCommandGroupRepo := new(MockCommandGroupRepository)
 		handler := handlers.NewAddCommandToGroupOnCommandDuplicated(mockCommandRepo, mockCommandGroupRepo)
 		event := commanddomainevent.CommandDuplicatedEvent{
@@ -161,7 +161,7 @@ func TestDefaultAddCommandToGroupOnCommandDuplicated(t *testing.T) {
 
 	t.Run("Should return error if failing to update command group", func(t *testing.T) {
 		// Arrange
-		mockCommandRepo := new(MockCommandRepository)
+		mockCommandRepo := new(test.MockCommandRepository)
 		mockCommandGroupRepo := new(MockCommandGroupRepository)
 		handler := handlers.NewAddCommandToGroupOnCommandDuplicated(mockCommandRepo, mockCommandGroupRepo)
 		event := commanddomainevent.CommandDuplicatedEvent{
