+++ internal/project/application/usecases/shared_test.go
@@ -5,7 +5,6 @@ import (
 
 	"github.com/stretchr/testify/mock"
 
-	commanddomain "gomander/internal/command/domain"
 	commandgroupdomain "gomander/internal/commandgroup/domain"
 	"gomander/internal/config/domain"
 	"gomander/internal/eventbus"
@@ -75,46 +74,6 @@ func (m *MockLogger) Debug(message string) {
 
 func (m *MockLogger) Error(message string) {
 	m.Called(message)
-}
-
-type MockCommandRepository struct {
-	mock.Mock
-}
-
-func (m *MockCommandRepository) Get(commandId string) (*commanddomain.Command, error) {
-	args := m.Called(commandId)
-	if args.Get(0) == nil {
-		return nil, args.Error(1)
-	}
-	return args.Get(0).(*commanddomain.Command), args.Error(1)
-}
-
-func (m *MockCommandRepository) GetAll(projectId string) ([]commanddomain.Command, error) {
-	args := m.Called(projectId)
-	if args.Get(0) == nil {
-		return nil, args.Error(1)
-	}
-	return args.Get(0).([]commanddomain.Command), args.Error(1)
-}
-
-func (m *MockCommandRepository) Create(command *commanddomain.Command) error {
-	args := m.Called(command)
-	return args.Error(0)
-}
-
-func (m *MockCommandRepository) Update(command *commanddomain.Command) error {
-	args := m.Called(command)
-	return args.Error(0)
-}
-
-func (m *MockCommandRepository) Delete(commandId string) error {
-	args := m.Called(commandId)
-	return args.Error(0)
-}
-
-func (m *MockCommandRepository) DeleteAll(projectId string) error {
-	args := m.Called(projectId)
-	return args.Error(0)
 }
 
 type MockFsFacade struct {
